requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-generativeai>=0.3.0
openai>=1.0.0
requests>=2.30.0
PyMuPDF==1.24.11
python-docx==0.8.11
Pillow>=10.0.0
reportlab==4.0.7
jinja2==3.1.2
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic>=2.0.0,<3.0.0
slowapi==0.1.8
